Rename Pipeline to LLMpipeline
